RecyclerView feeding data From Firebase
Updated  RecyclerViewAdapter to faciliate download of data from Firebase. Updated FragmentExpense to allow view of  data from Firebase. In Addition noted issue with Images - various sizes based on camera use. Changed Camera to 1:1 ratio which fixed issue but this is just a FIX and not a permanent solution to the problem. Will need to revisit compression to make sure height and width of compressed images in uniform. May also need to change the layout file for the Expenses tab in order to ensure movement of text does not occur. Noted issue with firebase data pulling into Expense Fragment with view ALL images and data stored. Have resolved this issue with databaseReference changed to expenses/users per my storage reference.