Modified Graph/Chart to use data from RecyclerView
NOTE - need to look at alternative as have discovered MPANDROIDCHART does NOT support RealTime Data...will have to review my code and see if I can find alternative.  Also data from Firebase to Graph is using a fix which I have had to make up through trial and error....at the moment each item in the array is populating the data in the PirChart. I will have to come up with a way of summarising the data based on the values in the String Array.