Update Graph to sum data by items in String Array
Pie Chart retrieves data from the recyclerview but now it takes the data and displays it as a summary or Expense types. i.e. accomodation, flights etc. refactored some methods, i.e. clear now is a method which is accessed by pressing both the clear button and the submit button....whereby once data is stored in firebase, the clear form method completes. This helps avoid issues with the user pressing the submit button twice and crashing the app.